refactor(layout): split default-layout/layout.tsx into per-composite modules
Replaces the single 817-LOC `src/skills/default-layout/layout.tsx`
file with a `layout/` directory of five focused submodules + a barrel
`index.ts`. Public surface is unchanged — every caller (`./layout`,
`../layout`) resolves through the new index.

Submodules:
- `mark.tsx`             — `AeMarkInline` brand monogram + `readUiScale`
- `grid.tsx`             — the CSS-grid `Layout` container with
                           `slotMap` remap + keep-mounted-for-motion
                           cells
- `status-bar.tsx`       — `StatusBar` (left/chip/center/right) and
                           the local `resolveValue` pointer helper
- `empty-state.tsx`      — `EmptyState` full-canvas welcome
- `worktree-landing.tsx` — `WorktreeLanding` + inner + the gh branch
                           status block

No behavior change. Each module's import surface is narrowed to what it
uses; the `mark` module is reused by `empty-state` and
`worktree-landing` so the AeMark stays single-source.
